fix: internal/localstorage разделить по файлам структуру
